pyproject.toml
[tool.poetry]
name = "task-tracker-api"
version = "0.1.0"
description = "Simple Flask API for task tracking with MongoDB"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.0"
flask-pymongo = "^2.3.0"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-flask = "^1.3.0"
pytest-cov = "^4.1.0"
mongomock = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
tdd_guard_project_root = "/Users/edward/PycharmProjects/TDD-guard-test"